feat: discovery queue
* feat: discovery queue in for you page

* feat: prettier discovery queue button

* refactor: queue data in profile controller

* chore: some more loggin
fix: for-you now properly needs a login to access

* Merge branch 'master' into feat/discovery-queue

# Conflicts:
#	persistence/src/main/java/ar/edu/itba/paw/persistence/UserDaoImpl.java
#	services-interfaces/src/main/java/ar/edu/itba/paw/servicesinterfaces/GameService.java

* fix: minor fixes and style issues

* Merge branch 'master' into feat/discovery-queue

* fix: one transactional left

* chore: remove some unused imports


Approved-by: Matías Manzur

+++ models/src/main/java/ar/edu/itba/paw/dtos/ReviewFilter.java
@@ -8,12 +8,12 @@ public class ReviewFilter {
 
     List<Integer> authorPreferences;
 
-    List<Integer> authors;
+    List<Long> authors;
 
     String reviewContent;
     Integer gameId;
 
-    public ReviewFilter(List<Integer> filterGameGenres, List<Integer> authorPreferences, List<Integer> authors, String reviewContent, Integer gameId) {
+    public ReviewFilter(List<Integer> filterGameGenres, List<Integer> authorPreferences, List<Long> authors, String reviewContent, Integer gameId) {
         this.filterGameGenres = filterGameGenres;
         this.authorPreferences = authorPreferences;
         this.authors = authors;
@@ -33,7 +33,7 @@ public class ReviewFilter {
         return authorPreferences;
     }
 
-    public List<Integer> getAuthors() {
+    public List<Long> getAuthors() {
         return authors;
     }
